Instructions and fixes for a 64-bit Windows build
This replaces COMPILING_WIN32.md with COMPILING_WIN64.md, containing
updated instructions for a 64-bit Windows build using Python 3.

It also adds a workaround for building CaDiCaL on Windows. CaDiCaL's
source tar contains a symlink that cannot be extracted on Windows. The
build script now ignores that file when extracting on Windows, it isn't
required for a successful build.

The resulting build hasn't been tested extensively, but it successfully
runs the included Python API examples.

Signed-off-by: Jannis Harder <me@jix.one>